requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
Automat==25.4.16
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
constantly==23.10.4
cryptography==44.0.2
cssselect==1.3.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
executing==2.2.0
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-speech==2.32.0
google-cloud-vision==3.10.1
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain-core==0.3.54
langgraph==0.3.31
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.32
lxml==5.3.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
openai==1.75.0
orjson==3.10.16
ormsgpack==1.9.1
outcome==1.3.0.post0
packaging==25.0
parsel==1.10.0
parso==0.8.4
pexpect==4.9.0
pinecone==6.0.2
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.7
prompt_toolkit==3.0.51
Protego==0.4.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyDispatcher==2.0.7
Pygments==2.19.1
pyOpenSSL==25.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
pyzmq==26.4.0
queuelib==1.8.0
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rsa==4.9.1
Scrapy==2.12.0
scrapy-selenium==0.0.7
selenium==4.31.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
tenacity==9.1.2
tldextract==5.2.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
Twisted==24.11.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
w3lib==2.3.1
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
xxhash==3.5.0
zope.interface==7.2
zstandard==0.23.0
